@@ -1,16 +1,17 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "io.techministry"
         minSdkVersion 15
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
-        buildConfigField "String", "API_KEY", '"' + "34833be0e1e5aa386acc219408a80c39" + '"'
+        buildConfigField "String", "API_URL_BIBLE", API_URL_BIBLE
+        buildConfigField "String", "API_KEY_BIBLE", API_KEY_BIBLE
     }
     buildTypes {
         release {
@@ -26,9 +27,9 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:support-v4:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0-beta01'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0-beta01'
     implementation "com.squareup.moshi:moshi:1.8.0"
     implementation 'com.squareup.retrofit2:converter-moshi:2.5.0'
     implementation 'com.squareup.retrofit2:adapter-rxjava2:2.5.0'
@@ -36,19 +37,37 @@ dependencies {
     implementation 'com.facebook.stetho:stetho:1.4.1'
     implementation 'com.facebook.stetho:stetho-okhttp3:1.4.1'
 
+    implementation 'com.squareup.okhttp3:logging-interceptor:3.11.0'
+    //Retrofit
+    implementation "com.squareup.retrofit2:adapter-rxjava:$rootProject.ext.rxJavaLibraryVersion"
+    implementation "com.squareup.retrofit2:retrofit:$rootProject.ext.rxJavaLibraryVersion"
+    implementation "com.squareup.retrofit2:converter-gson:$rootProject.ext.rxJavaLibraryVersion"
+    implementation "com.squareup.okhttp3:logging-interceptor:$rootProject.ext.okHttpVersion"
+    debugImplementation "com.squareup.leakcanary:leakcanary-android:$rootProject.ext.leakCanaryVersion"
+
+    //RxJava
+    implementation "io.reactivex:rxjava:$rootProject.ext.rxJavaVersion"
+    implementation "io.reactivex:rxandroid:$rootProject.ext.rxAndroidVersion"
+
     implementation 'com.nytimes.android:store3:3.1.0'
     implementation "com.nytimes.android:cache3:3.1.0"
     implementation "com.nytimes.android:middleware3:3.1.0"
     implementation "com.nytimes.android:filesystem3:3.1.0"
     implementation "com.google.code.gson:gson:2.8.2"
 
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
+    implementation 'com.google.android.material:material:1.0.0-beta01'
+    implementation 'androidx.cardview:cardview:1.0.0-beta01'
+    implementation 'androidx.appcompat:appcompat:1.0.0-beta01'
+    implementation 'com.google.android.material:material:1.0.0-beta01'
 
 
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test:runner:1.1.0-alpha4'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0-alpha4'
+
+    //Room
+    def room_version = "2.1.0-alpha07"
+
+    implementation "androidx.room:room-runtime:$room_version"
+    annotationProcessor "androidx.room:room-compiler:$room_version"
 }
