@@ -1,12 +1,11 @@
 package io.techministry.ui.bible;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,6 +16,7 @@ import io.techministry.network.BibleBook;
 public class BooksAdapter extends RecyclerView.Adapter<BooksAdapter.BookViewHolder> {
 
     private List<BibleBook> items;
+    private BooksListener mBooksListener;
 
     public BooksAdapter() {
         this.items = new ArrayList<>();
@@ -53,8 +53,25 @@ public class BooksAdapter extends RecyclerView.Adapter<BooksAdapter.BookViewHold
             ((TextView) itemView.findViewById(R.id.info_text)).setText(book.getName());
 //            itemView.findViewById(R.id.start);
 
-            itemView.setOnClickListener(view -> Toast.makeText(view.getContext(), "Clicked on " + book.getName(), Toast.LENGTH_LONG).show());
+            itemView.setOnClickListener(view -> {
+                if (mBooksListener != null)
+                    mBooksListener.onBookSelected(book);
+            });
+/*            itemView.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+
+                }
+            });*/
 //            itemView.setOnClickListener(View -> Toast.makeText(View.getContext(), "Clicked on " + book.getNameLong(), Toast.LENGTH_LONG).show());
         }
+    }
+
+    public void setBooksListener(BooksListener booksListener) {
+        mBooksListener = booksListener;
+    }
+
+    public interface BooksListener {
+        void onBookSelected(BibleBook book);
     }
 }
