@@ -8,12 +8,12 @@ public class BibleChapter {
       "reference": "Ruth 1"
       */
 
-    private String id;
-    private String bookId;
-    private int number;
-    private String reference;
+    String id;
+    String bookId;
+    String number;
+    String reference;
 
-    public BibleChapter(String id, String bookId, int number, String reference) {
+    public BibleChapter(String id, String bookId, String number, String reference) {
         this.id = id;
         this.bookId = bookId;
         this.number = number;
@@ -28,7 +28,7 @@ public class BibleChapter {
         return bookId;
     }
 
-    public int getNumber() {
+    public String getNumber() {
         return number;
     }
 
