@@ -1,12 +1,15 @@
 package io.techministry.network;
 
+import com.google.gson.annotations.SerializedName;
+
 import java.util.List;
 
 public class ChapterResponse {
+    @SerializedName("books")
     public List<BibleChapter> chapterList;
 
-    public ChapterResponse(List<BibleChapter> chapterList) {
-        this.chapterList = chapterList;
+    public ChapterResponse(List<BibleChapter> chList) {
+        this.chapterList = chList;
     }
 
     @Override
