@@ -1,23 +1,30 @@
 package io.techministry.ui;
 
 import android.os.Bundle;
-import android.support.design.widget.BottomNavigationView;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.widget.Toolbar;
+import android.view.MenuItem;
 
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
+import java.io.Serializable;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 import io.techministry.BaseActivity;
 import io.techministry.R;
+import io.techministry.network.BibleBook;
 import io.techministry.ui.bible.BibleFragment;
 import io.techministry.ui.bible.chapter.ChapterFragment;
 
-public class DashboardActivity extends BaseActivity {
+public class DashboardActivity extends BaseActivity implements BibleFragment.BibleFragmentDelegate {
 
     public BottomNavigationView bottomNavigationView;
 //    public Toolbar toolbar;
 
     private FragmentManager fragmentManager;
+    private BibleFragment mBibleFragment;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -25,15 +32,16 @@ public class DashboardActivity extends BaseActivity {
         setContentView(R.layout.activity_dashboard);
 
 //        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);  // Deprecated
+//        GodWorks heading title
         Toolbar toolbar = findViewById(R.id.toolbar);
+        setupBaseActionbar(toolbar, getString(R.string.app_name), false);
 
         fragmentManager = getSupportFragmentManager();
 
-        setupBaseActionbar(toolbar, getString(R.string.app_name), false);
 
         bottomNavigationView = findViewById(R.id.bottom_navigation);
         setBottomNavigation();
-
+        bottomNavigationView.setSelectedItemId(R.id.bottom_nav_bible);
     }
 
     @Override
@@ -49,47 +57,55 @@ public class DashboardActivity extends BaseActivity {
 //        bottomNavigationView.setSelectedItemId(R.id.bottom_nav_bible);
     }
 
-    private void setBottomNavigation() {
-        switchFragments(R.id.bottom_nav_bible);
+    @Override
+    public void onBackPressed() {
+        // TODO: show the appropriate fragment instead of closing the activity (if needed)
+        super.onBackPressed();
+    }
 
-        bottomNavigationView.setOnNavigationItemSelectedListener(item -> {
-            switchFragments(item.getItemId());
-            return true;
+    private void setBottomNavigation() {
+//        This is where the initial fragment is called/created; parameterized with int position
+//        switchFragments(R.id.bottom_nav_bible);
+
+        bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+                switchFragments(item.getItemId());
+                return true;
+            }
         });
     }
 
     private void switchFragments(int position) {
-        Fragment fragment = fragmentManager.findFragmentByTag(String.valueOf(position));
+        switchFragments(position, null);
+    }
+
+    private void switchFragments(int position, Serializable serializable) {
         FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+        Fragment newFrag = getItem(position, serializable);
+        fragmentTransaction.replace(R.id.container, newFrag, String.valueOf(position));
 
-        for (Fragment eachFragment : fragmentManager.getFragments()) {
-            fragmentTransaction.hide(eachFragment);
-        }
-
-        if (fragment == null) {
-            Fragment newFrag = getItem(position);
-            fragmentTransaction.add(R.id.container, newFrag, String.valueOf(position));
-            fragmentTransaction.commit();
-        } else {
-            fragmentTransaction.show(fragment);
-            fragmentTransaction.commit();
-        }
+        fragmentTransaction.commit();
     }
 
-    public Fragment getItem(int menu_id) {
-        Fragment fragment = null;
-
-//        fragment = new BibleFragment();
-
-
-        if (menu_id == R.id.bottom_nav_listen) {
-            fragment = new ChapterFragment();
-        }else if(menu_id == R.id.bottom_nav_bible) {
-            fragment = new BibleFragment();
-        }else if (menu_id == R.id.bottom_nav_the_word){
-            fragment = new BibleFragment();
+    // TODO : consider the Parcelable interface instead of Serializable
+    public Fragment getItem(int menu_id, Serializable serializable) {
+        switch (menu_id) {
+            case R.id.bottom_nav_listen:
+                BibleBook book = (BibleBook) serializable;
+                return ChapterFragment.getInstance(book);
+            case R.id.bottom_nav_bible:
+            case R.id.bottom_nav_the_word:
+                if (mBibleFragment == null)
+                    mBibleFragment = new BibleFragment();
+                return mBibleFragment;
         }
-        return fragment;
+
+        return null;
     }
 
+    @Override
+    public void onBookSelected(BibleBook book) {
+        switchFragments(R.id.bottom_nav_listen, book);
+    }
 }
